test(streaming): isolate early-exit lifecycle coverage
